Fix bug in NetworkHandler + minor tweaks

+++ lib/core_ext/integer.rb
@@ -1,4 +1,4 @@
-class Numeric
+class Integer
   #
   # Convert IP from number to String form
   #
